// src/external-services/airtable/airtable.types.ts
export interface AirtableOffer {
  OpportunityId: string;
  Entreprise: string;
  Titre: string;
  Nom: string;
  Prénom: string;
  Mail: string;
  'Contact mail': string;
  Téléphone: string;
  Fonction: string;
  'Description poste': string;
  'Description entreprise': string;
  'Compétences requises': string;
  'Pré-requis': string;
  "Secteur d'activité": string[];
  Publique: boolean;
  Externe: boolean;
  'Lien externe': string;
  'Origine externe': string;
  Validé: boolean;
  Archivé: boolean;
  'Date de création': string;
  Département: string;
  Adresse: string;
  Contrat: string;
  'Début de contrat': string;
  'Fin de contrat': string;
  'Temps partiel ?': boolean;
  'Nombre de postes': number;
  'Souhaite être recontacté': boolean;
  'Message personnalisé': string;
  'Permis de conduire': boolean;
  'Jours et horaires': string;
  Salaire: string;
  'Autres précisions': string;
  OpportunityUserId?: string;
  Candidat?: string;
  Statut?: string;
  Commentaire?: string;
  'Recommandée ?'?: boolean;
}